refactor: move display_decimal_256/128 to common-io.
