+++ src/common/io/Cargo.toml
@@ -23,6 +23,7 @@ chrono = { workspace = true }
 chrono-tz = { workspace = true }
 lexical-core = "0.8.5"
 micromarshal = "0.2.1"
+ethnum = "1.3.2"
 ordered-float = { workspace = true }
 serde = { workspace = true }
